+++ Runtime/Data/Int/IntReference.cs
@@ -1,0 +1,35 @@
+using System;
+
+namespace CaffeinatedStudios.ScriptableObjectArchitecture
+{
+    [Serializable]
+    public class IntReference : ReadOnlyIntReference
+    {
+        public IntReference()
+        { }
+
+        public IntReference(int value)
+        {
+            UseConstant = true;
+            ConstantValue = value;
+        }
+
+        public new int Value
+        {
+            get { return UseConstant ? ConstantValue : Variable.Value; }
+            set
+            {
+                if (UseConstant)
+                    ConstantValue = value;
+                else
+                    Variable.Value = value;
+            }
+        }
+
+        public static implicit operator int(IntReference reference)
+        {
+            return reference.Value;
+        }
+
+    }
+}