+++ Runtime/Data/Vector3/Vector3Reference.cs
@@ -1,0 +1,36 @@
+using System;
+using UnityEngine;
+
+namespace CaffeinatedStudios.ScriptableObjectArchitecture
+{
+
+    [Serializable]
+    public class Vector3Reference : ReadOnlyVector3Reference
+    {
+        public Vector3Reference()
+        { }
+
+        public Vector3Reference(Vector3 value)
+        {
+            UseConstant = true;
+            ConstantValue = value;
+        }
+
+        public new Vector3 Value
+        {
+            get { return UseConstant ? ConstantValue : Variable.Value; }
+            set
+            {
+                if (UseConstant)
+                    ConstantValue = value;
+                else
+                    Variable.SetValue(value);
+            }
+        }
+
+        public static implicit operator Vector3(Vector3Reference reference)
+        {
+            return reference.Value;
+        }
+    }
+}