+++ CHANGELOG.md
@@ -7,9 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed 
+
+- Improved the inspector for References
+
+
+## [1.0.2] - 2024-08-02
+
 ### Added
 
-- ReadOnly References for Float, Int, String and Vector2
+- ReadOnly References for Float, Int, String, Vector2, Vector3, and TimeSpan
+
+### Changed
+
+- Moved variables and references from Variables to Data folder
 
 
 ## [1.0.1] - 2024-08-01